Added 2 new api queries.
getUserPairsForTerm: gets pairs of users at least one of whom has sent a message to the other and both have used the search term in a tweet.
getSymPairsForTerm: gets pairs of users at least one of whome has sent a message to the other with the search term in it, both having used the search term.